Fix Color type replacing f32 by u8 for the type of the members

+++ src/color.rs
@@ -8,8 +8,8 @@ use crate::{
 
 #[derive(Default, Deserialize)]
 pub struct Color {
-    pub red: f32,
-    pub green: f32,
-    pub blue: f32,
-    pub alpha: f32,
+    pub red: u8,
+    pub green: u8,
+    pub blue: u8,
+    pub alpha: u8,
 }
